requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
httpie==2.4.0
idna==2.10
phonenumbers==8.12.22
psycopg2==2.8.6
Pygments==2.9.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.4
